🚚 rename: monthNames 파일명 MONTH_NAMES로 수정
상수는 scream case 사용하기
